Clear audit findings J/K/L: fractional-rate LTC (44100/29.97), open() resume guard, constant-time token
J (Med): LtcEncoder rewritten with phase-accurate fractional bit timing
(global-bit-anchored, drift-free); frameRate now Double. Proven by round-trip
tests at 44100/25, 44100/30, 48000/29.97 drop-frame. 23 pure-logic tests green.
K (Low): CameraController.open() guards all StateCallback paths with an
AtomicBoolean (also fixes a latent hang where onDisconnected during open never
resumed the coroutine).
L (Low): ExternalControlServer uses MessageDigest.isEqual for the pairing token
(constant-time) — verified compiling in the harness.
AUDIT_FINDINGS.md updated; remaining open items: I (UI-thread benchmark),
M (exported provider), N (WB math validation).

+++ packages/apps/OIWCamera/app/src/main/java/com/oiw/camera/audio/LtcDecoder.kt
@@ -13,10 +13,10 @@ package com.oiw.camera.audio
  */
 class LtcDecoder(
     sampleRate: Int = 48_000,
-    frameRate: Int = 25,
+    frameRate: Double = 25.0,
     private val onFrame: (LtcTimecode) -> Unit,
 ) {
-    private val samplesPerBit: Double = sampleRate.toDouble() / (frameRate * LtcTimecode.BITS_PER_FRAME)
+    private val samplesPerBit: Double = sampleRate / (frameRate * LtcTimecode.BITS_PER_FRAME)
     // Interval <= this is a half-bit ('1' pair); above is a full bit ('0'). Midway between 0.5 and 1.0.
     private val shortMax: Double = samplesPerBit * 0.75
     // Anything longer than this is a dropout/gap — reset the bit assembler rather than emit garbage.
